+++ pytorch/train_pets.py
@@ -21,26 +21,28 @@ import torchvision.transforms as T
 from torch.cuda.amp import autocast
 
 from utils import get_gpu_name
+from pets import get_pets_dataloader
 
-PROJECT = "pytorch-M1Pro"
+PROJECT = "pt2"
 ENTITY = "capecape"
 GROUP = "pytorch"
 
 config_defaults = SimpleNamespace(
-    batch_size=64,
-    device="mps",
+    batch_size=128,
+    device="cuda",
     epochs=1,
     num_experiments=1,
     learning_rate=1e-3,
-    image_size=128,
+    image_size=512,
     model_name="resnet50",
     dataset="PETS",
-    num_workers=4,
+    num_workers=0,
     gpu_name=get_gpu_name(),
     mixed_precision=False,
     channels_last=False,
     optimizer="Adam",
     compile=False,
+    tags=None,
 )
 
 def parse_args():
@@ -60,55 +62,8 @@ def parse_args():
     parser.add_argument('--channels_last', action="store_true")
     parser.add_argument('--optimizer', type=str, default=config_defaults.optimizer)
     parser.add_argument('--compile', action="store_true")
+    parser.add_argument('--tags', type=str, default=None)
     return parser.parse_args()
-
-def get_pets(version="v3"):
-    api = wandb.Api()
-    at = api.artifact(f'capecape/pytorch-M1Pro/PETS:{version}', type='dataset')
-    dataset_path = at.download()
-    return dataset_path
-
-class Pets(torch.utils.data.Dataset):
-    pat = r'(^[a-zA-Z]+_*[a-zA-Z]+)'
-    vocab = ['Abyssinian', 'Bengal', 'Birman', 'Bombay', 'British_Shorthair', 'Egyptian_Mau', 'Maine_Coon', 
-             'Persian', 'Ragdoll', 'Russian_Blue', 'Siamese', 'Sphynx', 'american_bulldog', 'american_pit', 
-             'basset_hound', 'beagle', 'boxer', 'chihuahua', 'english_cocker', 'english_setter', 'german_shorthaired', 
-             'great_pyrenees', 'havanese', 'japanese_chin', 'keeshond', 'leonberger', 'miniature_pinscher', 'newfoundland', 
-             'pomeranian', 'pug', 'saint_bernard', 'samoyed', 'scottish_terrier', 'shiba_inu', 'staffordshire_bull', 
-             'wheaten_terrier', 'yorkshire_terrier']
-    vocab_map = {v:i for i,v in enumerate(vocab)}
-
-
-    def __init__(self, pets_path, image_size=224):
-        self.path = Path(pets_path)
-        self.files = list(self.path.glob("images/*.jpg"))
-        self.tfms =T.Compose([T.Resize((image_size, image_size)), T.ToTensor()])
-        self.vocab_map = {v:i for i, v in enumerate(self.vocab)}
-    
-    @staticmethod
-    def load_image(fn, mode="RGB"):
-        "Open and load a `PIL.Image` and convert to `mode`"
-        im = Image.open(fn)
-        im.load()
-        im = im._new(im.im)
-        return im.convert(mode) if mode else im
-    
-    def __getitem__(self, idx):
-        file = self.files[idx]
-        return self.tfms(self.load_image(str(file))), self.vocab_map[re.match(self.pat, file.name)[0]]
-        
-    def __len__(self): return len(self.files)
-
-
-def get_dataloader(dataset_path, batch_size, image_size=224, num_workers=0, **kwargs):
-    "Get a training dataloader"
-    ds = Pets(dataset_path, image_size=image_size)
-    loader = torch.utils.data.DataLoader(ds, 
-                                         batch_size=batch_size,
-                                         pin_memory=True,
-                                         num_workers=num_workers,
-                                         **kwargs)
-    return loader
 
 
 def get_model(n_out, arch="resnet50"):
@@ -123,18 +78,19 @@ def check_cuda(config):
         config.mixed_precision = True
     return config
 
-def train(config=config_defaults):
+def train(config):
     config = check_cuda(config)
-    with wandb.init(project=PROJECT, entity=args.entity, group=GROUP, config=config):
+    tags = [f"pt{torch.__version__}", f"cuda{torch.version.cuda}"] + (config.tags.split(",") if config.tags is not None else [])
+    print(tags)
+    with wandb.init(project=PROJECT, entity=args.entity, group=GROUP, tags=tags, config=config):
 
         # Copy your config 
         config = wandb.config
 
         # Get the data
-        train_dl = get_dataloader(dataset_path=get_pets(), 
-                                  batch_size=config.batch_size, 
-                                  image_size=config.image_size,
-                                  num_workers=config.num_workers)
+        train_dl = get_pets_dataloader(batch_size=config.batch_size, 
+                                       image_size=config.image_size,
+                                       num_workers=config.num_workers)
         n_steps_per_epoch = math.ceil(len(train_dl.dataset) / config.batch_size)
 
         model = get_model(len(train_dl.dataset.vocab), config.model_name)
@@ -152,15 +108,15 @@ def train(config=config_defaults):
 
        # Training
         example_ct = 0
-        step_ct = 0
         for epoch in tqdm(range(config.epochs)):
-            t0 = perf_counter()
+            tf = t0 = perf_counter()
             model.train()
             for step, (images, labels) in enumerate(tqdm(train_dl, leave=False)):
                 images, labels = images.to(config.device), labels.to(config.device)
                 if config.channels_last:
                     images = images.contiguous(memory_format=torch.channels_last)
-
+                
+                # compute the model froward and backward pass time
                 ti = perf_counter()
                 if config.mixed_precision:
                     with autocast():
@@ -172,21 +128,23 @@ def train(config=config_defaults):
                 train_loss.backward()
                 optimizer.step()
                 optimizer.zero_grad()
+                tf_with_dataloader = perf_counter() - tf
                 tf = perf_counter()
+
+
+                # log the metrics
                 example_ct += len(images)
                 metrics = {"train/train_loss": train_loss, 
                            "train/epoch": (step + 1 + (n_steps_per_epoch * epoch)) / n_steps_per_epoch, 
                            "train/example_ct": example_ct,
                            "samples_per_sec":len(images)/(tf-ti),
+                           "samples_per_sec_dl":len(images)/tf_with_dataloader,
                            "samples_per_sec_epoch":example_ct/(tf-t0)}
 
                 if step + 1 < n_steps_per_epoch:
                     # 🐝 Log train metrics to wandb 
                     wandb.log(metrics)
 
-                step_ct += 1
-            
-                
 if __name__ == "__main__":
     args = parse_args()
     for _ in range(args.num_experiments):
